MAJOR Cleanup + UI enhancements
So everything works now. We have a fully functional exporter and a UI that isn't cancer by default. There's a few more things I need to add, but this is working great as of now. The convex hulls seem to be a little more accurate than anticipated, but hey, whatevs, we can fix that later.